requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paddleocr==2.9.1
paddlepaddle==2.6.2
pillow==10.4.0
httpx==0.27.2
python-multipart==0.0.9
pydantic-settings==2.5.2
